+++ src/view/RodarJanelas.java
@@ -1,0 +1,23 @@
+package view;
+
+import java.awt.EventQueue;
+
+public class RodarJanelas {
+
+	/**
+	 * Launch the application.
+	 */
+	public static void main(String[] args) {
+		EventQueue.invokeLater(new Runnable() {
+			public void run() {
+				try {
+					Teste frame = new Teste();
+					frame.setVisible(true);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
+	}
+
+}
